# contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: Next Clear Look Engine API
  version: 1.0.0
  summary: Sentinel-2 geometric opportunities, AOI archive evidence and historical likelihoods.
  description: |
    HTTP contract between the engine and the web app. An opportunity is geometry, not promised acquisition.
    A likelihood is empirical history, not a weather forecast. All instants are aware UTC RFC 3339.
servers:
  - url: http://localhost:8000/v1
    description: Local engine
...
paths:
...
  /mode:
    get:
      tags: [System]
      operationId: getMode
      summary: Read the active source mode and effective clock.
      responses:
        '200':
          description: Current mode.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mode'
              examples:
                replay:
                  externalValue: ./examples/mode-replay.json
                live:
                  externalValue: ./examples/mode-live.json
    put:
      tags: [System]
      operationId: switchMode
      summary: Switch source adapter and clock when no job is active.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModeSwitchRequest'
            examples:
              live:
                externalValue: ./examples/mode-switch-live.json
      responses:
        '200':
          description: Mode changed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mode'
              examples:
                live:
                  externalValue: ./examples/mode-live.json
        '409':
          description: Live mode is disabled, jobs are active, or the fixture set cannot be loaded.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                live_disabled:
                  externalValue: ./examples/error-live-disabled.json
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /events:
    get:
      tags: [Streams]
      operationId: streamLiveEvents
      summary: Resumeable mode clock, satellite-position, opportunity and new-scene events.
      parameters:
        - $ref: '#/components/parameters/LastEventIdHeader'
        - $ref: '#/components/parameters/LastEventIdQuery'
        - name: aoi_id
          in: query
          schema: {type: string}
          description: Optional AOI filter for opportunity/scene events.
      responses:
        '200':
          description: SSE live/replay stream; see events.md.
          content:
            text/event-stream:
              schema: {type: string}
              examples:
                replay:
                  externalValue: ./examples/live-events.sse
components:
  parameters:
...
    LastEventIdHeader:
      name: Last-Event-ID
      in: header
      required: false
      schema: {type: string, maxLength: 200}
    LastEventIdQuery:
      name: last_event_id
      in: query
      required: false
      schema: {type: string, maxLength: 200}
      description: Browser-compatible fallback when a client cannot set Last-Event-ID; header wins.
...
  responses:
...
    ValidationError:
      description: Request syntax was valid but one or more values failed validation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    Timestamp:
      type: string
      format: date-time
      examples: ['2026-10-02T17:43:54Z']
...
    Mode:
      type: object
      additionalProperties: false
      required: [mode, clock, network_enabled, live_available, fixture_set, fixture_recorded_at, deterministic]
      properties:
        mode: {type: string, enum: [live, replay]}
        clock: {$ref: '#/components/schemas/Timestamp'}
        network_enabled: {type: boolean}
        live_available: {type: boolean}
        fixture_set: {type: [string, 'null']}
        fixture_recorded_at:
          oneOf: [{$ref: '#/components/schemas/Timestamp'}, {type: 'null'}]
        deterministic: {type: boolean}
    ModeSwitchRequest:
      type: object
      additionalProperties: false
      required: [mode]
      properties:
        mode: {type: string, enum: [live, replay]}
        fixture_set: {type: string}
...
    ErrorObject:
      type: object
      additionalProperties: false
      required: [code, message, retryable, details, request_id]
      properties:
        code: {type: string, pattern: '^[A-Z][A-Z0-9_]+$'}
        message: {type: string}
        retryable: {type: boolean}
        details: {type: object, additionalProperties: true}
        request_id: {type: string}
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: {$ref: '#/components/schemas/ErrorObject'}